feat(leaderboard): displays username instead of First_Name
This commit fixes the issue where the code was incorrectly displaying the First_Name instead of the username in the user table. However, please note that there is still an unresolved bug related to the sorting of user data by username in certain scenarios. Further investigation is required to address this issue.